refactor(client): :recycle: Move Separator out into UI components
